Re-organized Super Editor tests for consistency (Resolves #703)
