+++ super_editor/test/super_editor/components/horizontal_rule_test.dart
@@ -5,7 +5,7 @@ import 'package:super_editor/src/test/super_editor_test/supereditor_robot.dart';
 import 'package:super_editor/super_editor.dart';
 
 import '../../test_tools.dart';
-import '../document_test_tools.dart';
+import '../supereditor_test_tools.dart';
 
 void main() {
   group('SuperEditor horizontal rule component', () {
