+++ super_editor/test/super_editor/components/list_items_test.dart
@@ -5,7 +5,7 @@ import 'package:super_editor/super_editor.dart';
 import 'package:super_editor/super_editor_test.dart';
 import 'package:super_text_layout/super_text_layout.dart';
 
-import '../document_test_tools.dart';
+import '../supereditor_test_tools.dart';
 import '../../test_tools.dart';
 
 void main() {
