+++ super_editor/test/super_editor/infrastructure/common_editor_operations_test.dart
@@ -2,10 +2,9 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:super_editor/super_editor.dart';
 import 'package:super_editor/super_editor_test.dart';
 
-import '../_document_test_tools.dart';
 import 'package:flutter_test_robots/flutter_test_robots.dart';
 
-import '../document_test_tools.dart';
+import '../supereditor_test_tools.dart';
 import '../../test_tools.dart';
 
 void main() {
