+++ super_editor/test/super_editor/infrastructure/editor_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:super_editor/super_editor.dart';
 
-import '../../super_editor/document_test_tools.dart';
+import '../../super_editor/supereditor_test_tools.dart';
 import '../../super_editor/test_documents.dart';
 
 void main() {
