+++ super_editor/test/super_editor/supereditor_components_test.dart
@@ -6,7 +6,7 @@ import 'package:super_editor/super_editor.dart';
 import 'package:super_editor/super_editor_test.dart';
 
 import '../test_tools.dart';
-import 'document_test_tools.dart';
+import 'supereditor_test_tools.dart';
 import 'test_documents.dart';
 
 void main() {
