+++ super_editor/test/super_editor/supereditor_copy_and_paste.dart
@@ -3,8 +3,8 @@ import 'package:flutter_test_robots/flutter_test_robots.dart';
 import 'package:super_editor/super_editor.dart';
 import 'package:super_editor/super_editor_test.dart';
 
-import '../../test_tools.dart';
-import '../document_test_tools.dart';
+import '../test_tools.dart';
+import 'supereditor_test_tools.dart';
 
 void main() {
   group('SuperEditor copy and paste > ', () {
