+++ super_editor/test/super_editor/supereditor_keyboard_test.dart
@@ -5,7 +5,7 @@ import 'package:super_editor/super_editor.dart';
 import 'package:super_editor/super_editor_test.dart';
 
 import '../test_tools.dart';
-import 'document_test_tools.dart';
+import 'supereditor_test_tools.dart';
 
 void main() {
   group('SuperEditor keyboard', () {
