+++ super_editor/test/super_editor/supereditor_software_keyboard_toolbar_test.dart
@@ -3,7 +3,7 @@ import 'package:super_editor/super_editor.dart';
 import 'package:super_editor/super_editor_test.dart';
 
 import '../test_tools.dart';
-import 'document_test_tools.dart';
+import 'supereditor_test_tools.dart';
 
 // TODO: Make the software keyboard toolbar testable
 //       Until then, this suite contains pieces of functionality that mirror the
