+++ super_editor/test/super_editor/text_entry/tagging/action_tags_test.dart
@@ -4,9 +4,9 @@ import 'package:flutter_test_robots/flutter_test_robots.dart';
 import 'package:super_editor/super_editor.dart';
 import 'package:super_editor/super_editor_test.dart';
 
-import '../../test_tools.dart';
-import '../document_test_tools.dart';
-import '../test_documents.dart';
+import '../../../test_tools.dart';
+import '../../supereditor_test_tools.dart';
+import '../../test_documents.dart';
 
 void main() {
   group("SuperEditor action tags >", () {
