+++ super_editor/test/super_editor/text_entry/tagging/pattern_tags_test.dart
@@ -3,9 +3,9 @@ import 'package:flutter_test_robots/flutter_test_robots.dart';
 import 'package:super_editor/super_editor.dart';
 import 'package:super_editor/super_editor_test.dart';
 
-import '../../test_tools.dart';
-import '../document_test_tools.dart';
-import '../test_documents.dart';
+import '../../../test_tools.dart';
+import '../../supereditor_test_tools.dart';
+import '../../test_documents.dart';
 
 void main() {
   group("SuperEditor pattern tags >", () {
