--- super_editor/test/test_tools_ime.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/services.dart';
-import 'package:super_editor/src/default_editor/document_ime/ime_decoration.dart';
-
-/// A [TextInputConnection] that tracks the number of content updates, to verify
-/// within tests.
-class ImeConnectionWithUpdateCount extends TextInputConnectionDecorator {
-  ImeConnectionWithUpdateCount(TextInputConnection client) : super(client);
-
-  int get contentUpdateCount => _contentUpdateCount;
-  int _contentUpdateCount = 0;
-
-  @override
-  void setEditingState(TextEditingValue value) {
-    super.setEditingState(value);
-    _contentUpdateCount += 1;
-  }
-}
