+++ super_editor/test/test_tools_user_input.dart
@@ -1,4 +1,50 @@
 import 'package:flutter/services.dart';
+import 'package:flutter_test/flutter_test.dart';
+import 'package:super_editor/super_editor.dart';
+
+final inputSourceVariant = ValueVariant({
+  TextInputSource.keyboard,
+  TextInputSource.ime,
+});
+
+final inputAndGestureVariants = ValueVariant<InputAndGestureTuple>(
+  {
+    const InputAndGestureTuple(TextInputSource.keyboard, DocumentGestureMode.mouse),
+    const InputAndGestureTuple(TextInputSource.keyboard, DocumentGestureMode.iOS),
+    const InputAndGestureTuple(TextInputSource.keyboard, DocumentGestureMode.android),
+    const InputAndGestureTuple(TextInputSource.ime, DocumentGestureMode.mouse),
+    const InputAndGestureTuple(TextInputSource.ime, DocumentGestureMode.iOS),
+    const InputAndGestureTuple(TextInputSource.ime, DocumentGestureMode.android),
+  },
+);
+
+/// A combination of an [inputSource] and a [gestureMode].
+class InputAndGestureTuple {
+  const InputAndGestureTuple(this.inputSource, this.gestureMode);
+
+  final TextInputSource inputSource;
+  final DocumentGestureMode gestureMode;
+
+  @override
+  String toString() {
+    return '${inputSource.name} Input Source & ${gestureMode.name} Gesture Mode';
+  }
+}
+
+/// A [TextInputConnection] that tracks the number of content updates, to verify
+/// within tests.
+class ImeConnectionWithUpdateCount extends TextInputConnectionDecorator {
+  ImeConnectionWithUpdateCount(TextInputConnection client) : super(client);
+
+  int get contentUpdateCount => _contentUpdateCount;
+  int _contentUpdateCount = 0;
+
+  @override
+  void setEditingState(TextEditingValue value) {
+    super.setEditingState(value);
+    _contentUpdateCount += 1;
+  }
+}
 
 /// Concrete version of [RawKeyEvent] used to manually simulate
 /// a specific key event sent from Flutter.
