+++ super_editor/test_goldens/editor/components/paragraph_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:golden_toolkit/golden_toolkit.dart';
 import 'package:super_editor/super_editor.dart';
 
-import '../../../test/super_editor/document_test_tools.dart';
+import '../../../test/super_editor/supereditor_test_tools.dart';
 import '../../test_tools_goldens.dart';
 
 void main() {
