+++ super_editor/test_goldens/editor/supereditor_text_layout_test.dart
@@ -5,7 +5,7 @@ import 'package:super_editor/src/test/super_editor_test/supereditor_inspector.da
 import 'package:super_editor/src/test/super_editor_test/supereditor_robot.dart';
 import 'package:super_editor/super_editor.dart';
 
-import '../../test/super_editor/document_test_tools.dart';
+import '../../test/super_editor/supereditor_test_tools.dart';
 import '../test_tools_goldens.dart';
 
 void main() {
